word generator
I wrote a(n inefficient) recursive algorithm that generates all possible combinations of English letters of a specified length n and determines if that combination is a word. I used a HashSet of strings as a database lol. I populated it with English words from this repo https://github.com/dwyl/english-words